Tests for label stats calculation
